cleaning up diff
