+++ src/lib/src/api/local/diff/utf8_diff.rs
@@ -1,35 +1,46 @@
 use crate::error::OxenError;
+use crate::model::diff::change_type::ChangeType;
+use crate::model::diff::text_diff::LineDiff;
+use crate::model::diff::text_diff::TextDiff;
 use crate::util;
 
-use colored::Colorize;
 use difference::{Changeset, Difference};
 use std::path::PathBuf;
 
-pub fn compare(version_file_1: &PathBuf, version_file_2: &PathBuf) -> Result<String, OxenError> {
+pub fn diff(version_file_1: &PathBuf, version_file_2: &PathBuf) -> Result<TextDiff, OxenError> {
     let original_data = util::fs::read_from_path(version_file_1)?;
     let compare_data = util::fs::read_from_path(version_file_2)?;
     let Changeset { diffs, .. } = Changeset::new(&original_data, &compare_data, "\n");
 
-    let mut outputs: Vec<String> = vec![];
+    let mut result = TextDiff { lines: vec![] };
     for diff in diffs {
         match diff {
             Difference::Same(ref x) => {
                 for split in x.split('\n') {
-                    outputs.push(format!(" {split}\n").normal().to_string());
+                    result.lines.push(LineDiff {
+                        modification: ChangeType::Unchanged,
+                        text: split.to_string(),
+                    });
                 }
             }
             Difference::Add(ref x) => {
                 for split in x.split('\n') {
-                    outputs.push(format!("+{split}\n").green().to_string());
+                    result.lines.push(LineDiff {
+                        modification: ChangeType::Added,
+                        text: split.to_string(),
+                    });
                 }
             }
             Difference::Rem(ref x) => {
                 for split in x.split('\n') {
-                    outputs.push(format!("-{split}\n").red().to_string());
+                    result.lines.push(LineDiff {
+                        modification: ChangeType::Removed,
+                        text: split.to_string(),
+                    });
                 }
             }
         }
     }
 
-    Ok(outputs.join(""))
+    Ok(result)
 }
