+++ src/lib/src/command.rs
@@ -35,7 +35,7 @@ pub use crate::command::checkout::{
 pub use crate::command::clone::{clone, clone_url, deep_clone_url, shallow_clone_url};
 pub use crate::command::commit::commit;
 pub use crate::command::df::{df, schema};
-pub use crate::command::diff::{diff_commits, diff_tabular};
+pub use crate::command::diff::{diff, diff_commits};
 pub use crate::command::fetch::fetch;
 pub use crate::command::info::info;
 pub use crate::command::init::init;
