+++ src/lib/src/model/diff.rs
@@ -1,11 +1,17 @@
 pub mod add_remove_modify_counts;
-pub use crate::model::diff::add_remove_modify_counts::AddRemoveModifyCounts;
+pub use add_remove_modify_counts::AddRemoveModifyCounts;
+
+pub mod change_type;
+pub use change_type::ChangeType;
 
 pub mod data_frame_diff;
 pub mod diff_commit_entry;
 pub mod diff_entries_counts;
 pub mod diff_entry;
 pub mod diff_entry_status;
+
+pub mod diff_result;
+pub use diff_result::DiffResult;
 
 pub mod generic_diff;
 pub mod generic_diff_summary;
