+++ src/lib/src/model/diff/change_type.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub enum ChangeType {
+    Added,
+    Removed,
+    Modified,
+    Unchanged,
+}
