+++ src/lib/src/model/diff/text_diff.rs
@@ -1,9 +1,13 @@
+use crate::model::diff::change_type::ChangeType;
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct LineDiff {
+    pub modification: ChangeType,
+    pub text: String,
+}
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct TextDiff {
-    pub added_lines: Vec<String>,
-    pub removed_lines: Vec<String>,
-    pub modified_lines: Vec<String>,
-    pub unchanged_lines: Vec<String>,
+    pub lines: Vec<LineDiff>,
 }
